# backend/openapi_milestone4.yaml
openapi: 3.0.3
info:
  title: PulseTrack HRMS API - Milestone 4
  version: 1.0.0
  description: |
    # PulseTrack HRMS - GenAI-Powered Human Resource Management System
    
    ## Project Information
    - **Team**: Team 11, IIT Madras
    - **Course**: Software Engineering (SEP 2025)
    - **Milestone**: 4 - API Endpoints Documentation
    
    ## Overview
    PulseTrack HRMS is a comprehensive, AI-powered Human Resource Management System designed to revolutionize 
    workforce management by making HR processes smarter, faster, and more human-centric.
    
    ## GenAI Integrations
    This system integrates multiple GenAI services to automate and enhance HR processes:
    
    ### 1. Google Gemini API
    - **Purpose**: Natural language processing, content generation, and intelligent analysis
    - **Used In**: Policy RAG, Resume Screening, Job Description Generation
    - **Model**: gemini-1.5-flash
    - **Features**: Text generation, embedding, multi-turn conversations
    
    ### 2. ChromaDB
    - **Purpose**: Vector database for semantic search
    - **Used In**: Policy RAG for document similarity search
    - **Features**: Persistent storage, similarity search, metadata filtering
    
    ### 3. Langchain
    - **Purpose**: RAG pipeline orchestration
    - **Used In**: Policy RAG question-answering system
    - **Features**: Document loading, text splitting, retrieval chains
    
    ### 4. PyPDF2
    - **Purpose**: PDF text extraction
...
    
...
    
...
    
...
    
...
    
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from /api/v1/auth/login endpoint"
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/employees:
    get:
      tags:
        - Employee Management
      summary: Get All Employees (Paginated)
      description: |
        **User Story #4: Employee Database**
        Maintain and update employee personal, job, and contact details in one central place.
        
        **Features:**
        - Pagination support
        - Search by name, email, or employee ID
        - Filter by department, team, role, active status
        - Returns summary view of employees
        
        **Error Handling:**
        - 403: HR access required
        - 422: Invalid query parameters
        
        **Access:** HR only
      operationId: getEmployees
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, email, or employee ID
        - name: department_id
          in: query
          schema:
            type: integer
        - name: team_id
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
            enum: [employee, manager, hr]
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  employees:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_pages:
                    type: integer
        '403':
          description: Forbidden - HR access required
    
    post:
      tags:
        - Employee Management
      summary: Create New Employee
      description: |
        **User Story #4: Employee Database**
        Create a new employee account with all required information.
        
        **Features:**
        - Auto-generate employee ID if not provided
        - Hash password securely
        - Set initial leave balances
        - Validate department, team, manager
        - Email uniqueness validation
        
        **Access:** HR only
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                position:
                  type: string
                department_id:
                  type: integer
                team_id:
                  type: integer
                manager_id:
                  type: integer
                role:
                  type: string
                  enum: [employee, manager, hr]
              required:
                - name
                - email
                - password
      responses:
        '201':
          description: Employee created successfully
        '400':
          description: Bad Request - Validation failed
        '403':
          description: Forbidden - HR access required
  /api/v1/employees/{employee_id}:
    get:
      tags:
        - Employee Management
      summary: Get Employee Details
      description: |
        Get complete employee details by ID.
        
        **Access:** HR only
      operationId: getEmployee
      parameters:
        - name: employee_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee details retrieved
        '404':
          description: Employee not found
    
    put:
      tags:
        - Employee Management
      summary: Update Employee
      description: |
        **User Story #4: Employee Database**
        Update employee details. All fields are optional.
        
        **Access:** HR only
      operationId: updateEmployee
      parameters:
        - name: employee_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                position:
                  type: string
                department_id:
                  type: integer
                is_active:
                  type: boolean
      responses:
        '200':
          description: Employee updated successfully
        '404':
          description: Employee not found
    
    delete:
      tags:
        - Employee Management
      summary: Deactivate Employee
      description: |
        **User Story #4: Employee Database**
        Soft delete employee (sets is_active to false).
        
        **Access:** HR only
      operationId: deactivateEmployee
      parameters:
        - name: employee_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee deactivated successfully
        '400':
          description: Cannot deactivate - has active team members
...